requirements.txt
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromedriver-py==111.0.5563.64
colorama==0.4.6
cryptography==40.0.1
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==23.0
pycparser==2.21
pyOpenSSL==23.1.1
PySocks==1.7.1
python-decouple==3.8
python-dotenv==0.21.1
requests==2.28.2
selenium==4.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.5.0
urllib3==1.26.15
urllib3-secure-extra==0.1.0
webdriver-manager==3.8.5
wsproto==1.2.0
